requirements.txt
pymongo==4.6.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytest==8.1.1